Object pool refactor
separated GenericObjectPool and ProjectileObjectPool to handle functionality individually
refactored GenericObjectPool to handle being a parent class to VFXObjectPool and CollectibleObjectPool (i.e. all non-ProjectileObjectPools)
updated VFXObjectPool to fit with GenericObjectPool; updated all affect classes that reference VFXObjectPool accordingly
created CollectibleObjectPool in the same format as VFXObjectPool
created (empty) Collectible class

+++ Assets/Scripts/Game/Actors/Projectiles/Projectile.cs
@@ -77,7 +77,7 @@ public abstract class Projectile : Actor
 
     protected void SpawnGrazeParticles(Vector3 spawnPos, float spawnRotZ)
     {
-        GameObject vfx = VFXObjectPool.Instance.Get(VFXType.PlayerGraze);
+        GameObject vfx = VFXObjectPool.Instance.Get((int)VFXType.PlayerGraze);
         var particleEffect = vfx.GetComponent<ParticleEffect>();
 
         //set spawn pos+rot
@@ -91,7 +91,7 @@ public abstract class Projectile : Actor
     public void SpawnDestructionParticles(Vector3 spawnPos)
     {
         //grab particle obj from pool; get VFX component
-        GameObject vfx = VFXObjectPool.Instance.Get(VFXType.BulletDestruction);
+        GameObject vfx = VFXObjectPool.Instance.Get((int)VFXType.BulletDestruction);
         var particleEffect = vfx.GetComponent<ParticleEffect>();
 
         vfx.transform.position = spawnPos;
