+++ Assets/Scripts/Game/Actors/Ship/Enemy.cs
@@ -200,7 +200,7 @@ public class Enemy : Ship
 
     public override void DisplayInvincibleShield(Vector3 spawnPos)
     {
-        GameObject vfx = VFXObjectPool.Instance.Get(VFXType.InvincibleEnemyShield);
+        GameObject vfx = VFXObjectPool.Instance.Get((int)VFXType.InvincibleEnemyShield);
         var particleEffect = vfx.GetComponent<ParticleEffect>();
 
         particleEffect.transform.position = transform.position;
