+++ Assets/Scripts/Game/VFX/Particle effects/PSBulletReflection.cs
@@ -1,10 +1,7 @@
-using System.Collections;
-using UnityEngine;
-
 public class PSBulletReflection : ParticleEffect
 {
     protected override void ReturnToPool()
     {
-        VFXObjectPool.Instance.ReturnToPool(gameObject, VFXType.BulletReflection);
+        VFXObjectPool.Instance.ReturnToPool(gameObject, (int)VFXType.BulletReflection);
     }
 }
