+++ Assets/Scripts/Game/VFX/Particle effects/PSInvincibleEnemyShield.cs
@@ -2,6 +2,6 @@ public class PSInvincibleEnemyShield : ParticleEffect
 {
     protected override void ReturnToPool()
     {
-        VFXObjectPool.Instance.ReturnToPool(gameObject, VFXType.InvincibleEnemyShield);
+        VFXObjectPool.Instance.ReturnToPool(gameObject, (int)VFXType.InvincibleEnemyShield);
     }
 }
