+++ Assets/Scripts/Game/VFX/Particle effects/PSPlayerGraze.cs
@@ -2,6 +2,6 @@ public class PSPlayerGraze : ParticleEffect
 {
     protected override void ReturnToPool()
     {
-        VFXObjectPool.Instance.ReturnToPool(gameObject, VFXType.PlayerGraze);
+        VFXObjectPool.Instance.ReturnToPool(gameObject, (int)VFXType.PlayerGraze);
     }
 }
